transport-university-chatbot/backend/requirements.txt
fastapi==0.121.1
uvicorn[standard]==0.38.0
SQLAlchemy==2.0.44
psycopg2-binary==2.9.11
pydantic==2.12.4
pydantic[email]==2.12.4
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-dotenv==1.2.1
alembic==1.17.1
pytest==9.0.0
httpx==0.28.1
starlette==0.49.3
